backend/requirements.txt
fastapi
uvicorn
pdf2image
requests
sentence-transformers
faiss-cpu
numpy